Remove adapter layer
