+++ index.html
@@ -79,8 +79,6 @@
   <!-- Module principal de l'application -->
   <script src="js/modules/app.js" defer></script>
 
-  <!-- Nouvelle entrée pour la migration -->
-  <script src="js/adapters/index.js" defer></script>
   
   <!-- Script d'initialisation des modules -->
   <script src="js/modules/init.js" defer></script>
